+++ pipfreeze/__init__.py
@@ -1,2 +1,2 @@
 from .pipfreeze import PipFreeze  # flake8: noqa
-from .requirement import Package, Requirement  # flake8: noqa
+from .package import Package, Requirement  # flake8: noqa