+++ CHANGES.rst
@@ -1,10 +1,16 @@
+Unreleased
+==========
+
+* implement ``Requirement``
+* implement ``Package``
+* implement ``PipFreeze.__contains__()``
+* implement ``PipFreeze.__len__()``
+
 0.0.2
 =====
 
 * implement ``PipFreeze.__bool__()`` (py3) and ``PipFreeze.__nonzero__()`` (py2)
 * implement ``PipFreeze.__iter__()``
-* implement ``Requirement``
-* implement ``Package``
 
 0.0.1
 =====